refactor: remove thin re-export files and fix eslint/prettier conflicts
Delete IdentityKeyManager.mjs and AddressKeyManager.mjs re-export shims,
point imports directly at sdkClient.mjs. Disable implicit-arrow-linebreak
and object-curly-newline eslint rules that conflict with prettier.
