--- tutorials/AddressKeyManager.mjs
@@ -1,1 +1,0 @@
-export { AddressKeyManager } from './sdkClient.mjs';
