--- tutorials/IdentityKeyManager.mjs
@@ -1,1 +1,0 @@
-export { IdentityKeyManager } from './sdkClient.mjs';
